+++ frontend/src/components/sidebar/General.vue
@@ -47,7 +47,7 @@
   </div>
 
   <!-- Section for logged-in users -->
-  <div v-if="loginCheck && !disableUsedPercentage" class="sidebar-scroll-list">
+  <div v-if="loginCheck" class="sidebar-scroll-list">
     <div class="sources card">
       <span>Sources</span>
       <div class="inner-card">
@@ -83,16 +83,8 @@
 
 <script>
 import * as auth from "@/utils/auth";
-import {
-  signup,
-  disableExternal,
-  disableUsedPercentage,
-  noAuth,
-  loginPage,
-} from "@/utils/constants";
-import { filesApi } from "@/api";
+import { signup, disableExternal, noAuth, loginPage } from "@/utils/constants";
 import ProgressBar from "@/components/ProgressBar.vue";
-import { getHumanReadableFilesize } from "@/utils/filesizes";
 import { state, getters, mutations } from "@/store"; // Import your custom store
 
 export default {
@@ -135,27 +127,7 @@ export default {
       }
     },
   },
-  mounted() {
-    if (this.loginCheck) {
-      this.updateUsage();
-    }
-  },
   methods: {
-    async updateUsage() {
-      if (!disableUsedPercentage) {
-        if (state.sources.info === undefined) {
-          return;
-        }
-        for (const source of Object.keys(state.sources.info)) {
-          let usage = await filesApi.usage(source);
-          let sourceInfo = state.sources.info[source];
-          sourceInfo.used = getHumanReadableFilesize(usage.used);
-          sourceInfo.total = getHumanReadableFilesize(usage.total);
-          sourceInfo.usedPercentage = Math.round((usage.used / usage.total) * 100);
-          mutations.updateSource(source, sourceInfo);
-        }
-      }
-    },
     checkLogin() {
       return getters.isLoggedIn() && !getters.routePath().startsWith("/share");
     },