+++ frontend/src/store/getters.js
@@ -1,10 +1,11 @@
 import { removePrefix } from "@/utils/url.js";
 import { getFileExtension } from  "@/utils/files.js";
 import { state } from "./state.js";
-import { mutations } from "./mutations.js";
-import { noAuth,baseURL } from "@/utils/constants.js";
+import { noAuth } from "@/utils/constants.js";
+import { getTypeInfo } from "@/utils/mimetype";
 
 export const getters = {
+  previewType: () => getTypeInfo(state.req.type).simpleType,
   isCardView: () => (state.user.viewMode == "gallery" || state.user.viewMode == "normal" ) && getters.currentView() == "listingView" ,
   currentHash: () => state.route.hash.replace("#", ""),
   isMobile: () => state.isMobile,
@@ -21,21 +22,8 @@ export const getters = {
     if (noAuth) {
       return true
     }
-    if (state.user !== null && state.user?.username != undefined && state.user?.username != "publicUser") {
+    if (state.user !== null && state.user?.username != "" && state.user?.username != "publicUser") {
       return true;
-    }
-    const userData = localStorage.getItem("userData");
-    if (userData == undefined) {
-      return false;
-    }
-    try {
-      const userInfo = JSON.parse(userData);
-      if (userInfo.username != "publicUser") {
-        mutations.setCurrentUser(userInfo);
-        return true;
-      }
-    } catch (error) {
-      return false;
     }
     return false
   },
@@ -94,34 +82,46 @@ export const getters = {
     return { dirs, files };
   },
   isSidebarVisible: () => {
+    const currentView = getters.currentView();
     let visible = (state.showSidebar || getters.isStickySidebar()) && state.user.username != "publicUser"
-    if (getters.currentView() == "settings") {
+    if (currentView == "settings") {
       visible = !getters.isMobile();
     }
-    if (getters.currentView() == "share") {
+    if (currentView == "share") {
       visible = false
     }
     if (typeof getters.currentPromptName() === "string" && !getters.isStickySidebar()) {
       visible = false;
     }
+    if (currentView == "preview") {
+      const contentType = getters.previewType()
+      if (
+        contentType == "audio" ||
+        contentType == "video" ||
+        contentType == "image" ||
+        contentType == "pdf"
+      ) {
+        visible = false
+      }
+    }
     if (
-        getters.currentView() == "editor" ||
-        getters.currentView() == "preview" ||
-        getters.currentView() == "onlyOfficeEditor"
+      currentView == "editor" ||
+      currentView == "onlyOfficeEditor"
       ) {
       visible = false;
     }
     return visible
   },
   isStickySidebar: () => {
-    let sticky = state.user?.stickySidebar
-    if (getters.currentView() == "settings") {
+    let sticky = state.user?.stickySidebar;
+    const currentView = getters.currentView();
+    if (currentView == "settings") {
       sticky = true
     }
-    if (getters.currentView() == null && !getters.isLoading() && getters.isShare() ) {
+    if (currentView == null && !getters.isLoading() && getters.isShare() ) {
       sticky = true
     }
-    if (getters.isMobile() || getters.currentView() == "preview") {
+    if (getters.isMobile()) {
       sticky = false
     }
     return sticky
@@ -142,7 +142,7 @@ export const getters = {
     // Step 1: Split the path by '/'
     let parts = urlPath.split("/");
     // Step 2: Assign hash to the second part (index 2) and join the rest for subPath
-    return baseURL + "share/"+ parts[1];
+    return "/share/"+ parts[1] + "/";
   },
   currentView: () => {
     const pathname = getters.routePath()