+++ frontend/src/utils/auth.js
@@ -1,9 +1,9 @@
-import { mutations, getters } from "@/store";
+import { mutations, getters, state } from "@/store";
 import router from "@/router";
-import { usersApi } from "@/api";
+import { usersApi,filesApi } from "@/api";
 import { getApiPath } from "@/utils/url.js";
-import { recaptcha, loginPage } from "@/utils/constants";
-
+import { recaptcha, loginPage, disableUsedPercentage } from "@/utils/constants";
+import { getHumanReadableFilesize } from "@/utils/filesizes";
 
 export async function setNewToken(token) {
   document.cookie = `auth=${token}; path=/`;
@@ -18,6 +18,17 @@ export async function validateLogin() {
   } catch (error) {
     console.log("Error validating login");
   }
+  if (!disableUsedPercentage) {
+    for (const source of Object.keys(state.sources.info)) {
+      let usage = await filesApi.usage(source);
+      let sourceInfo = state.sources.info[source];
+      sourceInfo.used = getHumanReadableFilesize(usage.used);
+      sourceInfo.total = getHumanReadableFilesize(usage.total);
+      sourceInfo.usedPercentage = Math.round((usage.used / usage.total) * 100);
+      mutations.updateSource(source, sourceInfo);
+    }
+  }
+
   return getters.isLoggedIn()
 }