+++ frontend/tests-noauth/preview.spec.ts
@@ -1,16 +1,18 @@
-import { test, expect } from "@playwright/test";
+import { test, expect } from "../test-setup";
 
-test("blob file preview", async ({  page, context }) => {
+test("blob file preview", async({ page, checkForErrors, context }) => {
   await page.goto("/files/");
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
   await page.locator('a[aria-label="file.tar.gz"]').waitFor({ state: 'visible' });
   await page.locator('a[aria-label="file.tar.gz"]').dblclick();
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - file.tar.gz");
   await page.locator('button[title="Close"]').click();
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files  - playwright-files");
+  await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
+  // Check for console errors
+  checkForErrors();
 });
 
-test("text file editor", async ({ page, context }) => {
+test("text file editor", async({ page, checkForErrors, context }) => {
   await page.goto("/files/");
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
   await page.locator('a[aria-label="copyme.txt"]').waitFor({ state: 'visible' });
@@ -20,9 +22,11 @@ test("text file editor", async ({ page, context }) => {
   expect(firstLineText).toBe('test file for playwright');
   await page.locator('button[title="Close"]').click();
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
+  // Check for console errors
+  checkForErrors();
 });
 
-test("navigate folders", async ({  page, context }) => {
+test("navigate folders", async({ page, checkForErrors, context }) => {
   await page.goto("/files/");
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - playwright-files");
   await page.locator('a[aria-label="myfolder"]').waitFor({ state: 'visible' });
@@ -34,21 +38,6 @@ test("navigate folders", async ({  page, context }) => {
   await page.locator('a[aria-label="gray-sample.jpg"]').waitFor({ state: 'visible' });
   await page.locator('a[aria-label="gray-sample.jpg"]').dblclick();
   await expect(page).toHaveTitle("Graham's Filebrowser - Files - gray-sample.jpg");
-});
-
-test("navigating images", async ({  page, context }) => {
-  await page.goto("/files/playwright-files/myfolder/testdata/20130612_142406.jpg");
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - 20130612_142406.jpg");
-  await page.locator('button[aria-label="Previous"]').waitFor({ state: 'hidden' });
-  await page.mouse.move(100, 100);
-  await page.locator('button[aria-label="Next"]').waitFor({ state: 'visible' });
-  await page.locator('button[aria-label="Next"]').click();
-  await expect(page).toHaveTitle("Graham's Filebrowser - Files - gray-sample.jpg");
-  await page.locator('button[aria-label="Previous"]').waitFor({ state: 'hidden' });
-  await page.locator('button[aria-label="Next"]').waitFor({ state: 'hidden' });
-  await page.mouse.move(100, 100);
-  await page.locator('button[aria-label="Next"]').waitFor({ state: 'visible' });
-  //await page.locator('button[aria-label="Next"]').click();
-  // went to next image
-  //await expect(page).toHaveTitle("Graham's Filebrowser - Files - IMG_2578.JPG");
+  // Check for console errors
+  checkForErrors();
 });